chore: refactor git/github api routes to server actions
