+++ apps/web/app/sessions/[sessionId]/chats/[chatId]/session-chat-content.tsx
@@ -45,8 +45,11 @@ import {
 import { createPortal } from "react-dom";
 import useSWR from "swr";
 import type { ChatRefreshResponse } from "@/app/api/sessions/[sessionId]/chats/[chatId]/route";
-import type { MergePullRequestResponse } from "@/app/api/sessions/[sessionId]/merge/route";
-import type { PrDeploymentResponse } from "@/app/api/sessions/[sessionId]/pr-deployment/route";
+import type { MergePullRequestResult } from "@/lib/github/actions/pr";
+import {
+  getDeploymentUrl,
+  type PrDeploymentResponse,
+} from "@/lib/github/queries/deployment";
 import type { CheckRun } from "@/lib/github/pulls";
 import type {
   WebAgentCommitDataPart,
@@ -119,7 +122,7 @@ import {
   estimateModelUsageCost,
 } from "@/lib/models";
 import { getPrDeploymentRefreshInterval } from "@/lib/pr-deployment-polling";
-import { fetcher } from "@/lib/swr";
+
 import { streamdownPlugins } from "@/lib/streamdown-config";
 import { cn } from "@/lib/utils";
 import {
@@ -2728,7 +2731,14 @@ export function SessionChatContent({
             prDeploymentQuery ? `?${prDeploymentQuery}` : ""
           }`
         : null,
-      fetcher,
+      async () =>
+        getDeploymentUrl({
+          sessionId: session.id,
+          ...(hasExistingPr && session.prNumber
+            ? { prNumber: session.prNumber }
+            : {}),
+          ...(previewLookupBranch ? { branch: previewLookupBranch } : {}),
+        }),
       {
         revalidateOnFocus: true,
         revalidateOnReconnect: true,
@@ -2870,7 +2880,7 @@ export function SessionChatContent({
   ]);
 
   const handleMerged = useCallback(
-    async (mergeResult: MergePullRequestResponse) => {
+    async (mergeResult: MergePullRequestResult) => {
       updateSessionPullRequest({
         prNumber: mergeResult.prNumber,
         prStatus: "merged",
